Move dark mode toggle down to second row on mobile
The full navbar buttons now have 0.5rem of horizontal padding instead
of 0.75rem as bootstrap has no class with that amount.
Additionally, The svgs for the toggle icons have been moved to css to
avoid having two copies of each in every html file. As a result, the
toggle icon no longer shrinks as the screen width does as bootstrap
has nothing like the .h3 class for width/height instead of font-size.